Module allocators::composable

This module contains some composable building blocks to build allocator chains.

Structs

Fallback
This allocator has a main and a fallback allocator. It will always attempt to allocate first with the main allocator, and second with the fallback.

NullAllocator
This allocator always fails. It will panic if you try to deallocate with it.

Proxy
This wraps an allocator and a logger, logging all allocations and deallocations.

Traits

ProxyLogger
Something that logs an allocator's activity. In practice, this may be an output stream, a data collector, or seomthing else entirely.