Trait allocators::BlockOwner

pub trait BlockOwner: Allocator {
    fn owns_block(&self, block: &Block) -> bool;

    fn owns<'a, T, A: Allocator>(&self, val: &AllocBox<'a, T, A>) -> bool { ... }
    fn with_fallback<O: BlockOwner>(self, other: O) -> Fallback<Self, O> where Self: Sized { ... }
}

An allocator that knows which blocks have been issued by it.

Required Methods

fn owns_block(&self, block: &Block) -> bool

Whether this allocator owns the block passed to it.

Provided Methods

fn owns<'a, T, A: Allocator>(&self, val: &AllocBox<'a, T, A>) -> bool

Whether this allocator owns this allocated value.

fn with_fallback<O: BlockOwner>(self, other: O) -> Fallback<Self, O> where Self: Sized

Joins this allocator with a fallback allocator.